feat: search with spec

+++ src/main/java/org/dainn/dainninventory/entity/ProductEntity.java
@@ -1,13 +1,12 @@
 package org.dainn.dainninventory.entity;
 
 import jakarta.persistence.*;
-import jakarta.validation.constraints.NotBlank;
-import jakarta.validation.constraints.NotNull;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.Setter;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
 
@@ -28,7 +27,7 @@ public class ProductEntity extends BaseEntity{
     private String description;
 
     @Column(name = "price", nullable = false)
-    private double price;
+    private BigDecimal price;
 
     @Column(name = "img_url", nullable = false)
     private String imgUrl;
